tests: update testing configuration to enable unix sockets to work during testing, somewhat
Due to issues with pytest_socket and pytest_vscode, debugging in vs code was not working.
Some claim this only affects WSL users. These changes at least make it so that a file can be debuggged in VS Code by using a launch task, but the test extension is still broken.

NOTE: this is not ideal because it completely enables sockets during tests, and not just the
local file socket. Eventually pytest_socket promised to provide a better implementation